Move training init function to trainer_main.py
